# User/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    Este API REST se encarga de gestionar la información y operaciones relacionadas con los usuarios dentro del microservicio de Usuarios del sistema de Netflix. Proporciona diversos endpoints para manejar el ciclo de vida completo de los usuarios, incluyendo la creación de cuentas, autenticación, actualización de perfiles, y la gestión de preferencias y configuraciones de cada usuario. Permite realizar operaciones CRUD (Crear, Leer, Actualizar, Eliminar) sobre los datos de usuario, asegurando que la información esté siempre segura y actualizada. Esta API facilita la interacción entre el frontend y el backend del sistema, garantizando que los usuarios puedan gestionar fácilmente sus cuentas y preferencias, mejorando la experiencia personalizada y segura dentro de la plataforma.
  title: API de Usuario
  version: "1.0"
servers:
- url: https://api.server/v1/user/
...
paths:
  /users:
    get:
      description: Obtiene todos los usuarios registrados en la plataforma.
      operationId: getUsers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: Operación exitosa
        "400":
          description: Solicitud inválida
        "404":
          description: Usuarios no encontrados
      summary: Obtiene todos los usuarios
      tags:
      - user
      x-accepts:
      - application/json
      x-tags:
      - tag: user
  /user/{userID}:
    delete:
      description: Elimina un usuario de la plataforma usando su ID.
      operationId: deleteUser
      parameters:
      - description: ID del usuario que se desea eliminar
        explode: false
        in: path
        name: userID
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/deleteUser_200_response'
          description: Usuario eliminado exitosamente
        "400":
          description: ID de usuario inválido proporcionado
        "404":
          description: Usuario no encontrado
      summary: Elimina un usuario existente
      tags:
      - user
      x-accepts:
      - application/json
      x-tags:
      - tag: user
    get:
      description: Obtiene la información de un usuario específico registrado en la
        plataforma utilizando su ID.
      operationId: getUserById
      parameters:
      - description: ID del usuario que se desea obtener
        explode: false
        in: path
        name: userID
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Operación exitosa
        "400":
          description: ID de usuario inválido proporcionado
        "404":
          description: Usuario no encontrado
      summary: Obtiene un usuario por su ID
      tags:
      - user
      x-accepts:
      - application/json
      x-tags:
      - tag: user
    put:
      description: Actualiza la información de un usuario existente en la plataforma
        por su ID.
      operationId: updateUser
      parameters:
      - description: ID del usuario que se desea actualizar
        explode: false
        in: path
        name: userID
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: Datos del usuario que se actualizarán
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Usuario actualizado con éxito
        "400":
          description: ID de usuario inválido proporcionado
        "404":
          description: Usuario no encontrado
        "422":
          description: Error de validación
      summary: Actualiza un usuario existente
      tags:
      - user
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: user
...
components:
  schemas:
    User:
      example:
        createdAt: 2023-10-09T10:20:30Z
        password: 1e34r
        preferences:
        - 1
        - 3
        - 4
        name: Juan Pérez
        userID: 1
        email: juan.perez@example.com
        username: juanperez
      properties:
        userID:
          description: ID único del usuario
          example: 1
          type: integer
        username:
          description: Nombre de usuario
          example: juanperez
          type: string
        password:
          description: Contraseña
          example: 1e34r
          type: string
        email:
          description: Dirección de correo electrónico del usuario
          example: juan.perez@example.com
          format: email
          type: string
        name:
          description: Nombre completo del usuario
          example: Juan Pérez
          type: string
        createdAt:
          description: Fecha y hora en que se creó el usuario
          example: 2023-10-09T10:20:30Z
          format: date-time
          type: string
        preferences:
          description: Array con los id's de géneros de contenido que le gustan al
            usuario
          example:
          - 1
          - 3
          - 4
          items:
            type: integer
          type: array
      required:
      - createdAt
      - email
      - name
      - preferences
      - userID
      - username
      type: object
...
    deleteUser_200_response:
      example:
        message: Usuario eliminado exitosamente.
      properties:
        message:
          example: Usuario eliminado exitosamente.
          type: string
      type: object
...